Better logs and errors and a bit of refactoring

+++ eznotes/default_editor.py
@@ -3,6 +3,8 @@ import readline
 
 from .const import DEFAULT_EDITOR_FILE_PATH
 from .utils import executable_exists
+from .logs import DefaultEditorLogs
+from .logs.error import executable_does_not_exist_error
 
 
 def editor_file_exists():
@@ -10,8 +12,12 @@ def editor_file_exists():
 
 
 def change_default_editor(new_editor):
-    with open(DEFAULT_EDITOR_FILE_PATH, 'w') as f:
-        f.write(new_editor)
+    if executable_exists(new_editor):
+        with open(DEFAULT_EDITOR_FILE_PATH, 'w') as f:
+            f.write(new_editor)
+    else:
+        executable_does_not_exist_error(new_editor)
+
 
 
 def get_default_editor():
@@ -26,12 +32,14 @@ def editor_initiate():
     # also gets called.
     if editor_file_exists():
         return
+
+    logs = DefaultEditorLogs()
+    logs.first()
+
     editor_exists = False
     while not editor_exists:
-        print(
-            "Please enter the name of your default editor (it can be changed later)"
-        )
-        new_editor = input("Editor (leave blank for 'vim'): ")
+        logs.input_prompt()
+        new_editor = input()
         if new_editor == '':
             new_editor = 'vim'
         editor_exists = executable_exists(new_editor)
